@@ -12,10 +12,11 @@ using DSharpPlus.Entities;
 using DSharpPlus.Entities.AuditLogs;
 using DSharpPlus.EventArgs;
 using DSharpPlus.Net.Abstractions;
+using DSharpPlus.Net.InboundWebhooks;
+using DSharpPlus.Net.InboundWebhooks.Payloads;
 using DSharpPlus.Net.Serialization;
 
 using Microsoft.Extensions.Logging;
-
 using Newtonsoft.Json.Linq;
 
 namespace DSharpPlus;
@@ -80,11 +81,8 @@ public sealed partial class DiscordClient
             #region Gateway Status
 
             case "ready":
-                JArray? glds = (JArray)dat["guilds"];
-                JArray? dmcs = (JArray)dat["private_channels"];
-
-                dat.Remove("guilds");
-                dat.Remove("private_channels");
+                JArray? glds = (JArray?)dat["guilds"];
+                JArray? dmcs = (JArray?)dat["private_channels"];
 
                 int readyShardId = payload is ShardIdContainingGatewayPayload { ShardId: { } id } ? id : 0;
 
@@ -582,21 +580,102 @@ public sealed partial class DiscordClient
                 await OnAutoModerationRuleExecutedAsync(dat.ToDiscordObject<DiscordAutoModerationActionExecution>());
                 break;
                 #endregion
-            
+
             #region Entitlements
             case "entitlement_create":
                 await OnEntitlementCreatedAsync(dat.ToDiscordObject<DiscordEntitlement>());
                 break;
-            
+
             case "entitlement_update":
                 await OnEntitlementUpdatedAsync(dat.ToDiscordObject<DiscordEntitlement>());
                 break;
-            
+
             case "entitlement_delete":
                 await OnEntitlementDeletedAsync(dat.ToDiscordObject<DiscordEntitlement>());
                 break;
             #endregion
         }
+    }
+
+    #endregion
+
+    #region Webhook Events
+
+    private async Task ReceiveWebhookEventsAsync()
+    {
+        while (!this.webhookEventReader.Completion.IsCompleted)
+        {
+            DiscordWebhookEvent payload = await this.webhookEventReader.ReadAsync();
+
+            try
+            {
+                await HandleWebhookDispatchAsync(payload);
+            }
+            catch (Exception ex)
+            {
+                this.Logger.LogError(ex, "Dispatch threw an exception: ");
+            }
+        }
+    }
+
+    private async Task ReceiveInteractionEventsAsync()
+    {
+        while (!this.interactionEventReader.Completion.IsCompleted)
+        {
+            DiscordHttpInteractionPayload payload = await this.interactionEventReader.ReadAsync();
+            DiscordHttpInteraction interaction = payload.ProtoInteraction;
+
+            ulong? guildId = interaction.GuildId;
+            ulong channelId = interaction.ChannelId;
+
+            JToken rawMember = payload.Data["member"];
+            TransportMember? transportMember = null;
+            TransportUser transportUser;
+            if (rawMember != null)
+            {
+                transportMember = payload.Data["member"].ToDiscordObject<TransportMember>();
+                transportUser = transportMember.User;
+            }
+            else
+            {
+                transportUser = payload.Data["user"].ToDiscordObject<TransportUser>();
+            }
+
+            DiscordChannel channel = interaction.Channel;
+            channel.Discord = this;
+
+            await OnInteractionCreateAsync(guildId, channelId, transportUser, transportMember, channel, interaction);
+        }
+    }
+
+    private Task HandleWebhookDispatchAsync(DiscordWebhookEvent @event)
+    {
+        if (@event.ApplicationID != this.CurrentApplication.Id)
+        {
+            this.Logger.LogCritical
+            (
+                "The application event webhook received an event for application {OtherId}, which is different from the current application.",
+                @event.ApplicationID
+            );
+
+            return Task.CompletedTask;
+        }
+
+        if (@event.Type == DiscordWebhookEventType.Ping)
+        {
+            return Task.CompletedTask;
+        }
+
+        DiscordWebhookEventBody body = @event.Event;
+
+        _ = body.Type switch
+        {
+            DiscordWebhookEventBodyType.ApplicationAuthorized => OnApplicationAuthorizedAsync(body),
+            DiscordWebhookEventBodyType.EntitlementCreate => OnWebhookEntitlementCreateAsync(body),
+            _ => OnUnknownWebhookEventAsync(body)
+        };
+
+        return Task.CompletedTask;
     }
 
     #endregion
@@ -1569,6 +1648,8 @@ public sealed partial class DiscordClient
             ISet<ulong> nf = dat["not_found"].ToDiscordObject<ISet<ulong>>();
             ea.NotFound = new ReadOnlySet<ulong>(nf);
         }
+
+        _ = this.DispatchGuildMembersChunkForIteratorsAsync(ea);
 
         await this.dispatcher.DispatchAsync(this, ea);
     }
@@ -2630,6 +2711,46 @@ public sealed partial class DiscordClient
 
     #region Misc
 
+    internal async Task OnApplicationAuthorizedAsync(DiscordWebhookEventBody body)
+    {
+        ApplicationAuthorizedPayload payload = body.Data.ToDiscordObject<ApplicationAuthorizedPayload>();
+
+        DiscordGuild? guild = payload.Guild;
+        DiscordUser user = payload.User;
+
+        UpdateUserCache(user);
+
+        if (guild is not null)
+        {
+            if (this.guilds.TryGetValue(guild.Id, out DiscordGuild? cachedGuild))
+            {
+                guild = cachedGuild;
+            }
+
+            guild.Discord = this;
+
+            if (guild.Members.TryGetValue(user.Id, out DiscordMember? member))
+            {
+                user = member;
+            }
+        }
+        else
+        {
+            user.Discord = this;
+        }
+
+        ApplicationAuthorizedEventArgs eventArgs = new()
+        {
+            Guild = guild,
+            IntegrationType = payload.IntegrationType,
+            Scopes = payload.Scopes,
+            User = user,
+            Timestamp = body.Timestamp
+        };
+
+        await this.dispatcher.DispatchAsync(this, eventArgs);
+    }
+
     internal async Task OnInteractionCreateAsync(ulong? guildId, ulong channelId, TransportUser user, TransportMember member, DiscordChannel? channel, DiscordInteraction interaction)
     {
         DiscordUser usr = new(user) { Discord = this };
@@ -2639,7 +2760,7 @@ public sealed partial class DiscordClient
         interaction.Discord = this;
         interaction.Data.Discord = this;
 
-        if (member != null)
+        if (member is not null && guildId is not null && interaction.Guild is not null)
         {
             usr = new DiscordMember(member) { guild_id = guildId.Value, Discord = this };
             UpdateUser(usr, guildId, interaction.Guild, member);
@@ -2837,6 +2958,17 @@ public sealed partial class DiscordClient
         await this.dispatcher.DispatchAsync(this, ea);
     }
 
+    internal async Task OnUnknownWebhookEventAsync(DiscordWebhookEventBody body)
+    {
+        UnknownEventArgs eventArgs = new()
+        {
+            EventName = body.Type.ToString(),
+            Json = body.Data?.ToString()
+        };
+
+        await this.dispatcher.DispatchAsync(this, eventArgs);
+    }
+
     #endregion
 
     #region AutoModeration
@@ -2877,14 +3009,27 @@ public sealed partial class DiscordClient
     #endregion
 
     #region Entitlements
-    
-    private async Task OnEntitlementCreatedAsync(DiscordEntitlement entitlement) 
+
+    private async Task OnEntitlementCreatedAsync(DiscordEntitlement entitlement)
         => await this.dispatcher.DispatchAsync(this, new EntitlementCreatedEventArgs { Entitlement = entitlement });
 
-    private async Task OnEntitlementUpdatedAsync(DiscordEntitlement entitlement) 
+    private async Task OnWebhookEntitlementCreateAsync(DiscordWebhookEventBody body)
+    {
+        await this.dispatcher.DispatchAsync
+        (
+            this,
+            new EntitlementCreatedEventArgs
+            {
+                Entitlement = body.Data.ToDiscordObject<DiscordEntitlement>(),
+                Timestamp = body.Timestamp
+            }
+        );
+    }
+
+    private async Task OnEntitlementUpdatedAsync(DiscordEntitlement entitlement)
         => await this.dispatcher.DispatchAsync(this, new EntitlementUpdatedEventArgs { Entitlement = entitlement });
 
-    private async Task OnEntitlementDeletedAsync(DiscordEntitlement entitlement) 
+    private async Task OnEntitlementDeletedAsync(DiscordEntitlement entitlement)
         => await this.dispatcher.DispatchAsync(this,  new EntitlementDeletedEventArgs { Entitlement = entitlement });
 
     #endregion
