@@ -6,11 +6,11 @@ namespace DSharpPlus.Commands.ContextChecks;
 [AttributeUsage(AttributeTargets.Class | AttributeTargets.Method | AttributeTargets.Delegate)]
 public class RequirePermissionsAttribute : RequireGuildAttribute
 {
-    public DiscordPermissions BotPermissions { get; init; }
-    public DiscordPermissions UserPermissions { get; init; }
+    public DiscordPermission[] BotPermissions { get; init; }
+    public DiscordPermission[] UserPermissions { get; init; }
 
-    public RequirePermissionsAttribute(DiscordPermissions permissions) => this.BotPermissions = this.UserPermissions = permissions;
-    public RequirePermissionsAttribute(DiscordPermissions botPermissions, DiscordPermissions userPermissions)
+    public RequirePermissionsAttribute(params DiscordPermission[] permissions) => this.BotPermissions = this.UserPermissions = permissions;
+    public RequirePermissionsAttribute(DiscordPermission[] botPermissions, DiscordPermission[] userPermissions)
     {
         this.BotPermissions = botPermissions;
         this.UserPermissions = userPermissions;
