@@ -11,9 +11,12 @@ internal sealed class RequirePermissionsCheck : IContextCheck<RequirePermissions
 {
     public ValueTask<string?> ExecuteCheckAsync(RequirePermissionsAttribute attribute, CommandContext context)
     {
+        DiscordPermissions requiredBotPermissions = new(attribute.BotPermissions);
+        DiscordPermissions requiredUserPermissions = new(attribute.UserPermissions);
+
         if (context is SlashCommandContext slashContext)
         {
-            if (!slashContext.Interaction.AppPermissions.HasPermission(attribute.BotPermissions))
+            if (!slashContext.Interaction.AppPermissions.HasAllPermissions(requiredBotPermissions))
             {
                 return ValueTask.FromResult<string?>("The bot did not have the needed permissions to execute this command.");
             }
@@ -24,11 +27,11 @@ internal sealed class RequirePermissionsCheck : IContextCheck<RequirePermissions
         {
             return ValueTask.FromResult<string?>(RequireGuildCheck.ErrorMessage);
         }
-        else if (!context.Guild!.CurrentMember.PermissionsIn(context.Channel).HasFlag(attribute.BotPermissions))
+        else if (!context.Guild!.CurrentMember.PermissionsIn(context.Channel).HasAllPermissions(requiredBotPermissions))
         {
             return ValueTask.FromResult<string?>("The bot did not have the needed permissions to execute this command.");
         }
-        else if (!context.Member!.PermissionsIn(context.Channel).HasFlag(attribute.UserPermissions))
+        else if (!context.Member!.PermissionsIn(context.Channel).HasAllPermissions(requiredUserPermissions))
         {
             return ValueTask.FromResult<string?>("The executing user did not have the needed permissions to execute this command.");
         }
