@@ -65,7 +65,7 @@ public sealed class MessageCommandProcessor : ICommandProcessor
                 MessageCommandLogging.messageCommandCannotHaveSubcommands(logger, command.FullName, null);
                 continue;
             }
-            else if (!command.Method!.GetParameters()[0].ParameterType.IsAssignableFrom(typeof(SlashCommandContext)))
+            else if (!command.Method!.GetParameters()[0].ParameterType.IsAssignableFrom(typeof(MessageCommandContext)))
             {
                 MessageCommandLogging.messageCommandContextParameterType(logger, command.FullName, null);
                 continue;
@@ -177,13 +177,18 @@ public sealed class MessageCommandProcessor : ICommandProcessor
             nameLocalizations = await this.slashCommandProcessor.ExecuteLocalizerAsync(localizerAttribute.LocalizerType, $"{command.FullName}.name");
         }
 
-        return new(
+        DiscordPermission[]? userPermissions = command.Attributes.OfType<RequirePermissionsAttribute>().FirstOrDefault()?.UserPermissions;
+
+        return new
+        (
             name: command.Attributes.OfType<DisplayNameAttribute>().FirstOrDefault()?.DisplayName ?? command.FullName,
             description: string.Empty,
             type: DiscordApplicationCommandType.MessageContextMenu,
             name_localizations: nameLocalizations,
             allowDMUsage: command.Attributes.Any(x => x is AllowDMUsageAttribute),
-            defaultMemberPermissions: command.Attributes.OfType<RequirePermissionsAttribute>().FirstOrDefault()?.UserPermissions ?? DiscordPermissions.UseApplicationCommands,
+            defaultMemberPermissions: userPermissions is not null
+                ? new(userPermissions)
+                : new DiscordPermissions(DiscordPermission.UseApplicationCommands),
             nsfw: command.Attributes.Any(x => x is RequireNsfwAttribute),
             contexts: command.Attributes.OfType<InteractionAllowedContextsAttribute>().FirstOrDefault()?.AllowedContexts,
             integrationTypes: command.Attributes.OfType<InteractionInstallTypeAttribute>().FirstOrDefault()?.InstallTypes
