@@ -1,7 +1,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using DSharpPlus.Commands.Converters;
-using DSharpPlus.Commands.Processors.SlashCommands.NamingPolicies;
+using DSharpPlus.Commands.Processors.SlashCommands.InteractionNamingPolicies;
 using DSharpPlus.Entities;
 
 namespace DSharpPlus.Commands.Processors.SlashCommands;
