@@ -50,29 +50,41 @@ public sealed class DefaultEventDispatcher : IEventDispatcher, IDisposable
         IReadOnlyList<object> general = this.handlers[typeof(DiscordEventArgs)];
         IReadOnlyList<object> specific = this.handlers[typeof(T)];
 
-        IServiceScope scope = this.serviceProvider.CreateScope();
-
         if (general.Count == 0 && specific.Count == 0)
         {
             return ValueTask.CompletedTask;
         }
 
-        _ = Task.WhenAll
-        (
-            general.Concat(specific)
-                .Select(async handler =>
-                {
-                    try
-                    {
-                        await ((Func<DiscordClient, T, IServiceProvider, Task>)handler)(client, eventArgs, scope.ServiceProvider);
-                    }
-                    catch (Exception e)
-                    {
-                        await this.errorHandler.HandleEventHandlerError(typeof(T).ToString(), e, (Delegate)handler, client, eventArgs);
-                    }
-                })
-        )
-        .ContinueWith((_) => scope.Dispose());
+        try
+        {
+            IServiceScope scope = this.serviceProvider.CreateScope();
+            _ = Task.WhenAll
+                (
+                    general.Concat(specific)
+                        .Select(async handler =>
+                        {
+                            try
+                            {
+                                await ((Func<DiscordClient, T, IServiceProvider, Task>)handler)(client, eventArgs,
+                                    scope.ServiceProvider);
+                            }
+                            catch (Exception e)
+                            {
+                                await this.errorHandler.HandleEventHandlerError(typeof(T).ToString(), e,
+                                    (Delegate)handler, client, eventArgs);
+                            }
+                        })
+                )
+                .ContinueWith((_) => scope.Dispose());
+        }
+        catch (ObjectDisposedException)
+        {
+            // ObjectDisposedException can be thrown from the this.serviceProvider.CreateScope() call above,
+            // when the serviceProvider is already disposed externally.
+            // This *should* only happen when the hosting application is shutting down,
+            // so it should be safe to just ignore it.
+            // One option would be to show that exception as debug log, but I would guess that just causes confusion.
+        }
 
         return ValueTask.CompletedTask;
     }
