@@ -120,8 +120,9 @@ public class DiscordChannel : SnowflakeObject, IEquatable<DiscordChannel>
     public int? UserLimit { get; internal set; }
 
     /// <summary>
-    /// <para>Gets the slow mode delay configured for this channel.</para>
-    /// <para>All bots, as well as users with <see cref="DiscordPermissions.ManageChannels"/> or <see cref="DiscordPermissions.ManageMessages"/> permissions in the channel are exempt from slow mode.</para>
+    /// Gets the slow mode delay configured for this channel.<br/>
+    /// All bots, as well as users with <see cref="DiscordPermission.ManageChannels"/> 
+    /// or <see cref="DiscordPermission.ManageMessages"/> permissions in the channel are exempt from slow mode.
     /// </summary>
     [JsonProperty("rate_limit_per_user")]
     public int? PerUserRateLimit { get; internal set; }
@@ -149,27 +150,45 @@ public class DiscordChannel : SnowflakeObject, IEquatable<DiscordChannel>
     /// Gets this channel's children. This applies only to channel categories.
     /// </summary>
     [JsonIgnore]
-    public IReadOnlyList<DiscordChannel> Children => !this.IsCategory
+    public IReadOnlyList<DiscordChannel> Children
+    {
+        get
+        {
+            return !this.IsCategory
                 ? throw new ArgumentException("Only channel categories contain children.")
                 : this.Guild.channels.Values.Where(e => e.ParentId == this.Id).ToList();
+        }
+    }
 
     /// <summary>
     /// Gets this channel's threads. This applies only to text and news channels.
     /// </summary>
     [JsonIgnore]
-    public IReadOnlyList<DiscordThreadChannel> Threads => this.Type is not (DiscordChannelType.Text or DiscordChannelType.News or DiscordChannelType.GuildForum)
+    public IReadOnlyList<DiscordThreadChannel> Threads
+    {
+        get
+        {
+            return this.Type is not (DiscordChannelType.Text or DiscordChannelType.News or DiscordChannelType.GuildForum)
                 ? throw new ArgumentException("Only text channels can have threads.")
                 : this.Guild.threads.Values.Where(e => e.ParentId == this.Id).ToArray();
+        }
+    }
 
     /// <summary>
     /// Gets the list of members currently in the channel (if voice channel), or members who can see the channel (otherwise).
     /// </summary>
     [JsonIgnore]
-    public virtual IReadOnlyList<DiscordMember> Users => this.Guild is null
+    public virtual IReadOnlyList<DiscordMember> Users
+    {
+        get
+        {
+            return this.Guild is null
                 ? throw new InvalidOperationException("Cannot query users outside of guild channels.")
                 : (IReadOnlyList<DiscordMember>)(this.Type is DiscordChannelType.Voice or DiscordChannelType.Stage
                 ? this.Guild.Members.Values.Where(x => x.VoiceState?.ChannelId == this.Id).ToList()
-                : this.Guild.Members.Values.Where(x => (PermissionsFor(x) & DiscordPermissions.AccessChannels) == DiscordPermissions.AccessChannels).ToList());
+                : this.Guild.Members.Values.Where(x => PermissionsFor(x).HasPermission(DiscordPermission.ViewChannel)).ToList());
+        }
+    }
 
     /// <summary>
     /// Gets whether this channel is an NSFW channel.
@@ -203,7 +222,9 @@ public class DiscordChannel : SnowflakeObject, IEquatable<DiscordChannel>
     /// </summary>
     /// <param name="content">Content of the message to send.</param>
     /// <returns>The sent message.</returns>
-    /// <exception cref="UnauthorizedException">Thrown when the client does not have the <see cref="DiscordPermissions.SendMessages"/> permission if TTS is true and <see cref="DiscordPermissions.SendTtsMessages"/> if TTS is true.</exception>
+    /// <exception cref="UnauthorizedException">Thrown when the client does not have the 
+    /// <see cref="DiscordPermission.SendMessages"/> permission if TTS is false and 
+    /// <see cref="DiscordPermission.SendTtsMessages"/> if TTS is true.</exception>
     /// <exception cref="NotFoundException">Thrown when the channel does not exist.</exception>
     /// <exception cref="BadRequestException">Thrown when an invalid parameter was provided.</exception>
     /// <exception cref="ServerErrorException">Thrown when Discord is unable to process the request.</exception>
@@ -216,7 +237,9 @@ public class DiscordChannel : SnowflakeObject, IEquatable<DiscordChannel>
     /// </summary>
     /// <param name="embed">Embed to attach to the message.</param>
     /// <returns>The sent message.</returns>
-    /// <exception cref="UnauthorizedException">Thrown when the client does not have the <see cref="DiscordPermissions.SendMessages"/> permission if TTS is true and <see cref="DiscordPermissions.SendTtsMessages"/> if TTS is true.</exception>
+    /// <exception cref="UnauthorizedException">Thrown when the client does not have the 
+    /// <see cref="DiscordPermission.SendMessages"/> permission if TTS is false and 
+    /// <see cref="DiscordPermission.SendTtsMessages"/> if TTS is true.</exception>
     /// <exception cref="NotFoundException">Thrown when the channel does not exist.</exception>
     /// <exception cref="BadRequestException">Thrown when an invalid parameter was provided.</exception>
     /// <exception cref="ServerErrorException">Thrown when Discord is unable to process the request.</exception>
@@ -230,7 +253,9 @@ public class DiscordChannel : SnowflakeObject, IEquatable<DiscordChannel>
     /// <param name="embed">Embed to attach to the message.</param>
     /// <param name="content">Content of the message to send.</param>
     /// <returns>The sent message.</returns>
-    /// <exception cref="UnauthorizedException">Thrown when the client does not have the <see cref="DiscordPermissions.SendMessages"/> permission if TTS is true and <see cref="DiscordPermissions.SendTtsMessages"/> if TTS is true.</exception>
+    /// <exception cref="UnauthorizedException">Thrown when the client does not have the 
+    /// <see cref="DiscordPermission.SendMessages"/> permission if TTS is false and 
+    /// <see cref="DiscordPermission.SendTtsMessages"/> if TTS is true.</exception>
     /// <exception cref="NotFoundException">Thrown when the channel does not exist.</exception>
     /// <exception cref="BadRequestException">Thrown when an invalid parameter was provided.</exception>
     /// <exception cref="ServerErrorException">Thrown when Discord is unable to process the request.</exception>
@@ -243,7 +268,9 @@ public class DiscordChannel : SnowflakeObject, IEquatable<DiscordChannel>
     /// </summary>
     /// <param name="builder">The builder with all the items to send.</param>
     /// <returns>The sent message.</returns>
-    /// <exception cref="UnauthorizedException">Thrown when the client does not have the <see cref="DiscordPermissions.SendMessages"/> permission TTS is true and <see cref="DiscordPermissions.SendTtsMessages"/> if TTS is true.</exception>
+    /// <exception cref="UnauthorizedException">Thrown when the client does not have the 
+    /// <see cref="DiscordPermission.SendMessages"/> permission if TTS is false and 
+    /// <see cref="DiscordPermission.SendTtsMessages"/> if TTS is true.</exception>
     /// <exception cref="NotFoundException">Thrown when the channel does not exist.</exception>
     /// <exception cref="BadRequestException">Thrown when an invalid parameter was provided.</exception>
     /// <exception cref="ServerErrorException">Thrown when Discord is unable to process the request.</exception>
@@ -256,7 +283,9 @@ public class DiscordChannel : SnowflakeObject, IEquatable<DiscordChannel>
     /// </summary>
     /// <param name="action">The builder with all the items to send.</param>
     /// <returns>The sent message.</returns>
-    /// <exception cref="UnauthorizedException">Thrown when the client does not have the <see cref="DiscordPermissions.SendMessages"/> permission TTS is true and <see cref="DiscordPermissions.SendTtsMessages"/> if TTS is true.</exception>
+    /// <exception cref="UnauthorizedException">Thrown when the client does not have the 
+    /// <see cref="DiscordPermission.SendMessages"/> permission if TTS is false and 
+    /// <see cref="DiscordPermission.SendTtsMessages"/> if TTS is true.</exception>
     /// <exception cref="NotFoundException">Thrown when the channel does not exist.</exception>
     /// <exception cref="BadRequestException">Thrown when an invalid parameter was provided.</exception>
     /// <exception cref="ServerErrorException">Thrown when Discord is unable to process the request.</exception>
@@ -294,7 +323,7 @@ public class DiscordChannel : SnowflakeObject, IEquatable<DiscordChannel>
     /// </summary>
     /// <param name="reason">Reason for audit logs.</param>
     /// <returns></returns>
-    /// <exception cref="UnauthorizedException">Thrown when the client does not have the <see cref="DiscordPermissions.ManageChannels"/> permission.</exception>
+    /// <exception cref="UnauthorizedException">Thrown when the client does not have the <see cref="DiscordPermission.ManageChannels"/> permission.</exception>
     /// <exception cref="NotFoundException">Thrown when the channel does not exist.</exception>
     /// <exception cref="BadRequestException">Thrown when an invalid parameter was provided.</exception>
     /// <exception cref="ServerErrorException">Thrown when Discord is unable to process the request.</exception>
@@ -306,7 +335,7 @@ public class DiscordChannel : SnowflakeObject, IEquatable<DiscordChannel>
     /// </summary>
     /// <param name="reason">Reason for audit logs.</param>
     /// <returns>Newly-created channel.</returns>
-    /// <exception cref="UnauthorizedException">Thrown when the client does not have the <see cref="DiscordPermissions.ManageChannels"/> permission.</exception>
+    /// <exception cref="UnauthorizedException">Thrown when the client does not have the <see cref="DiscordPermission.ManageChannels"/> permission.</exception>
     /// <exception cref="NotFoundException">Thrown when the channel does not exist.</exception>
     /// <exception cref="BadRequestException">Thrown when an invalid parameter was provided.</exception>
     /// <exception cref="ServerErrorException">Thrown when Discord is unable to process the request.</exception>
@@ -343,7 +372,7 @@ public class DiscordChannel : SnowflakeObject, IEquatable<DiscordChannel>
     /// <param name="id">The ID of the message</param>
     /// <param name="skipCache">Whether to always make a REST request.</param>
     /// <returns></returns>
-    /// <exception cref="UnauthorizedException">Thrown when the client does not have the <see cref="DiscordPermissions.ReadMessageHistory"/> permission.</exception>
+    /// <exception cref="UnauthorizedException">Thrown when the client does not have the <see cref="DiscordPermission.ReadMessageHistory"/> permission.</exception>
     /// <exception cref="NotFoundException">Thrown when the channel does not exist.</exception>
     /// <exception cref="BadRequestException">Thrown when an invalid parameter was provided.</exception>
     /// <exception cref="ServerErrorException">Thrown when Discord is unable to process the request.</exception>
@@ -359,7 +388,7 @@ public class DiscordChannel : SnowflakeObject, IEquatable<DiscordChannel>
     /// </summary>
     /// <param name="action">Action to perform on this channel</param>
     /// <returns></returns>
-    /// <exception cref="UnauthorizedException">Thrown when the client does not have the <see cref="DiscordPermissions.ManageChannels"/> permission.</exception>
+    /// <exception cref="UnauthorizedException">Thrown when the client does not have the <see cref="DiscordPermission.ManageChannels"/> permission.</exception>
     /// <exception cref="NotFoundException">Thrown when the channel does not exist.</exception>
     /// <exception cref="BadRequestException">Thrown when an invalid parameter was provided.</exception>
     /// <exception cref="ServerErrorException">Thrown when Discord is unable to process the request.</exception>
@@ -401,7 +430,7 @@ public class DiscordChannel : SnowflakeObject, IEquatable<DiscordChannel>
     /// <param name="lockPermissions">Whether to sync channel permissions with the parent, if moving to a new category.</param>
     /// <param name="parentId">The new parent ID if the channel is to be moved to a new category.</param>
     /// <returns></returns>
-    /// <exception cref="UnauthorizedException">Thrown when the client does not have the <see cref="DiscordPermissions.ManageChannels"/> permission.</exception>
+    /// <exception cref="UnauthorizedException">Thrown when the client does not have the <see cref="DiscordPermission.ManageChannels"/> permission.</exception>
     /// <exception cref="NotFoundException">Thrown when the channel does not exist.</exception>
     /// <exception cref="BadRequestException">Thrown when an invalid parameter was provided.</exception>
     /// <exception cref="ServerErrorException">Thrown when Discord is unable to process the request.</exception>
@@ -434,7 +463,8 @@ public class DiscordChannel : SnowflakeObject, IEquatable<DiscordChannel>
     /// <param name="before">Message to fetch before from.</param>
     /// <param name="cancellationToken">Cancels the enumeration before doing the next api request</param>
     /// </summary>
-    /// <exception cref="UnauthorizedException">Thrown when the client does not have the <see cref="DiscordPermissions.AccessChannels"/> permission.</exception>
+    /// <exception cref="UnauthorizedException">Thrown when the client does not have the 
+    /// <see cref="DiscordPermission.ViewChannel"/> permission.</exception>
     /// <exception cref="NotFoundException">Thrown when the channel does not exist.</exception>
     /// <exception cref="BadRequestException">Thrown when an invalid parameter was provided.</exception>
     /// <exception cref="ServerErrorException">Thrown when Discord is unable to process the request.</exception>
@@ -447,7 +477,8 @@ public class DiscordChannel : SnowflakeObject, IEquatable<DiscordChannel>
     /// <param name="after">Message to fetch after from.</param>
     /// <param name="cancellationToken">Cancels the enumeration before doing the next api request</param>
     /// </summary>
-    /// <exception cref="UnauthorizedException">Thrown when the client does not have the <see cref="DiscordPermissions.AccessChannels"/> permission.</exception>
+    /// <exception cref="UnauthorizedException">Thrown when the client does not have the 
+    /// <see cref="DiscordPermission.ViewChannel"/> permission.</exception>
     /// <exception cref="NotFoundException">Thrown when the channel does not exist.</exception>
     /// <exception cref="BadRequestException">Thrown when an invalid parameter was provided.</exception>
     /// <exception cref="ServerErrorException">Thrown when Discord is unable to process the request.</exception>
@@ -460,7 +491,8 @@ public class DiscordChannel : SnowflakeObject, IEquatable<DiscordChannel>
     /// <param name="around">Message to fetch around from.</param>
     /// <param name="cancellationToken">Cancels the enumeration before doing the next api request</param>
     /// </summary>
-    /// <exception cref="UnauthorizedException">Thrown when the client does not have the <see cref="DiscordPermissions.AccessChannels"/> permission.</exception>
+    /// <exception cref="UnauthorizedException">Thrown when the client does not have the 
+    /// <see cref="DiscordPermission.ViewChannel"/> permission.</exception>
     /// <exception cref="NotFoundException">Thrown when the channel does not exist.</exception>
     /// <exception cref="BadRequestException">Thrown when an invalid parameter was provided.</exception>
     /// <exception cref="ServerErrorException">Thrown when Discord is unable to process the request.</exception>
@@ -472,7 +504,8 @@ public class DiscordChannel : SnowflakeObject, IEquatable<DiscordChannel>
     /// <param name="limit">The amount of messages to fetch.</param>
     /// <param name="cancellationToken">Cancels the enumeration before doing the next api request</param>
     /// </summary>
-    /// <exception cref="UnauthorizedException">Thrown when the client does not have the <see cref="DiscordPermissions.AccessChannels"/> permission.</exception>
+    /// <exception cref="UnauthorizedException">Thrown when the client does not have the 
+    /// <see cref="DiscordPermission.ViewChannel"/> permission.</exception>
     /// <exception cref="NotFoundException">Thrown when the channel does not exist.</exception>
     /// <exception cref="BadRequestException">Thrown when an invalid parameter was provided.</exception>
     /// <exception cref="ServerErrorException">Thrown when Discord is unable to process the request.</exception>
@@ -539,6 +572,7 @@ public class DiscordChannel : SnowflakeObject, IEquatable<DiscordChannel>
                 {
                     yield return msg;
                 }
+
                 last = sortedMessageArray.LastOrDefault()?.Id;
             }
             else
@@ -547,6 +581,7 @@ public class DiscordChannel : SnowflakeObject, IEquatable<DiscordChannel>
                 {
                     yield return sortedMessageArray[i];
                 }
+
                 last = sortedMessageArray.FirstOrDefault()?.Id;
             }
         }
@@ -557,7 +592,7 @@ public class DiscordChannel : SnowflakeObject, IEquatable<DiscordChannel>
     /// Gets the threads that are public and archived for this channel.
     /// </summary>
     /// <returns>A <seealso cref="ThreadQueryResult"/> containing the threads for this query and if an other call will yield more threads.</returns>
-    /// <exception cref="UnauthorizedException">Thrown when the client does not have the <see cref="DiscordPermissions.ReadMessageHistory"/> permission.</exception>
+    /// <exception cref="UnauthorizedException">Thrown when the client does not have the <see cref="DiscordPermission.ReadMessageHistory"/> permission.</exception>
     /// <exception cref="NotFoundException">Thrown when the channel does not exist.</exception>
     /// <exception cref="BadRequestException">Thrown when an invalid parameter was provided.</exception>
     /// <exception cref="ServerErrorException">Thrown when Discord is unable to process the request.</exception>
@@ -569,7 +604,9 @@ public class DiscordChannel : SnowflakeObject, IEquatable<DiscordChannel>
     /// Gets the threads that are private and archived for this channel.
     /// </summary>
     /// <returns>A <seealso cref="ThreadQueryResult"/> containing the threads for this query and if an other call will yield more threads.</returns>
-    /// <exception cref="UnauthorizedException">Thrown when the client does not have the <see cref="DiscordPermissions.ReadMessageHistory"/> and the <see cref="DiscordPermissions.ManageThreads"/> permission.</exception>
+    /// <exception cref="UnauthorizedException">Thrown when the client does not have the 
+    /// <see cref="DiscordPermission.ReadMessageHistory"/> and the 
+    /// <see cref="DiscordPermission.ManageThreads"/> permission.</exception>
     /// <exception cref="NotFoundException">Thrown when the channel does not exist.</exception>
     /// <exception cref="BadRequestException">Thrown when an invalid parameter was provided.</exception>
     /// <exception cref="ServerErrorException">Thrown when Discord is unable to process the request.</exception>
@@ -581,7 +618,8 @@ public class DiscordChannel : SnowflakeObject, IEquatable<DiscordChannel>
     /// Gets the private and archived threads that the current member has joined in this channel.
     /// </summary>
     /// <returns>A <seealso cref="ThreadQueryResult"/> containing the threads for this query and if an other call will yield more threads.</returns>
-    /// <exception cref="UnauthorizedException">Thrown when the client does not have the <see cref="DiscordPermissions.ReadMessageHistory"/> permission.</exception>
+    /// <exception cref="UnauthorizedException">Thrown when the client does not have the 
+    /// <see cref="DiscordPermission.ReadMessageHistory"/> permission.</exception>
     /// <exception cref="NotFoundException">Thrown when the channel does not exist.</exception>
     /// <exception cref="BadRequestException">Thrown when an invalid parameter was provided.</exception>
     /// <exception cref="ServerErrorException">Thrown when Discord is unable to process the request.</exception>
@@ -632,6 +670,7 @@ public class DiscordChannel : SnowflakeObject, IEquatable<DiscordChannel>
                         throw new ArgumentException("You can only delete messages that are less than 14 days old.");
                     }
                 }
+
                 await this.Discord.ApiClient.DeleteMessagesAsync(this.Id,
                     messageBatch.Select(x => x.Id), reason);
                 deleteCount += takeCount;
@@ -697,7 +736,8 @@ public class DiscordChannel : SnowflakeObject, IEquatable<DiscordChannel>
     /// <param name="message">The message to be deleted.</param>
     /// <param name="reason">Reason for audit logs.</param>
     /// <returns></returns>
-    /// <exception cref="UnauthorizedException">Thrown when the client does not have the <see cref="DiscordPermissions.ManageMessages"/> permission.</exception>
+    /// <exception cref="UnauthorizedException">Thrown when the client does not have the 
+    /// <see cref="DiscordPermission.ManageMessages"/> permission.</exception>
     /// <exception cref="NotFoundException">Thrown when the channel does not exist.</exception>
     /// <exception cref="BadRequestException">Thrown when an invalid parameter was provided.</exception>
     /// <exception cref="ServerErrorException">Thrown when Discord is unable to process the request.</exception>
@@ -708,7 +748,8 @@ public class DiscordChannel : SnowflakeObject, IEquatable<DiscordChannel>
     /// Returns a list of invite objects
     /// </summary>
     /// <returns></returns>
-    /// <exception cref="UnauthorizedException">Thrown when the client does not have the <see cref="DiscordPermissions.CreateInstantInvite"/> permission.</exception>
+    /// <exception cref="UnauthorizedException">Thrown when the client does not have the 
+    /// <see cref="DiscordPermission.CreateInvite"/> permission.</exception>
     /// <exception cref="NotFoundException">Thrown when the channel does not exist.</exception>
     /// <exception cref="BadRequestException">Thrown when an invalid parameter was provided.</exception>
     /// <exception cref="ServerErrorException">Thrown when Discord is unable to process the request.</exception>
@@ -728,7 +769,8 @@ public class DiscordChannel : SnowflakeObject, IEquatable<DiscordChannel>
     /// <param name="targetUserId">The ID of the target user.</param>
     /// <param name="targetApplicationId">The ID of the target application.</param>
     /// <returns></returns>
-    /// <exception cref="UnauthorizedException">Thrown when the client does not have the <see cref="DiscordPermissions.CreateInstantInvite"/> permission.</exception>
+    /// <exception cref="UnauthorizedException">Thrown when the client does not have the 
+    /// <see cref="DiscordPermission.CreateInvite"/> permission.</exception>
     /// <exception cref="NotFoundException">Thrown when the channel does not exist.</exception>
     /// <exception cref="BadRequestException">Thrown when an invalid parameter was provided.</exception>
     /// <exception cref="ServerErrorException">Thrown when Discord is unable to process the request.</exception>
@@ -743,11 +785,11 @@ public class DiscordChannel : SnowflakeObject, IEquatable<DiscordChannel>
     /// <param name="deny">The permissions to deny.</param>
     /// <param name="reason">Reason for audit logs.</param>
     /// <returns></returns>
-    /// <exception cref="UnauthorizedException">Thrown when the client does not have the <see cref="DiscordPermissions.ManageRoles"/> permission.</exception>
+    /// <exception cref="UnauthorizedException">Thrown when the client does not have the <see cref="DiscordPermission.ManageRoles"/> permission.</exception>
     /// <exception cref="NotFoundException">Thrown when the channel does not exist.</exception>
     /// <exception cref="BadRequestException">Thrown when an invalid parameter was provided.</exception>
     /// <exception cref="ServerErrorException">Thrown when Discord is unable to process the request.</exception>
-    public async Task AddOverwriteAsync(DiscordMember member, DiscordPermissions allow = DiscordPermissions.None, DiscordPermissions deny = DiscordPermissions.None, string? reason = null)
+    public async Task AddOverwriteAsync(DiscordMember member, DiscordPermissions allow = default, DiscordPermissions deny = default, string? reason = null)
         => await this.Discord.ApiClient.EditChannelPermissionsAsync(this.Id, member.Id, allow, deny, "member", reason);
 
     /// <summary>
@@ -758,11 +800,11 @@ public class DiscordChannel : SnowflakeObject, IEquatable<DiscordChannel>
     /// <param name="deny">The permissions to deny.</param>
     /// <param name="reason">Reason for audit logs.</param>
     /// <returns></returns>
-    /// <exception cref="UnauthorizedException">Thrown when the client does not have the <see cref="DiscordPermissions.ManageRoles"/> permission.</exception>
+    /// <exception cref="UnauthorizedException">Thrown when the client does not have the <see cref="DiscordPermission.ManageRoles"/> permission.</exception>
     /// <exception cref="NotFoundException">Thrown when the channel does not exist.</exception>
     /// <exception cref="BadRequestException">Thrown when an invalid parameter was provided.</exception>
     /// <exception cref="ServerErrorException">Thrown when Discord is unable to process the request.</exception>
-    public async Task AddOverwriteAsync(DiscordRole role, DiscordPermissions allow = DiscordPermissions.None, DiscordPermissions deny = DiscordPermissions.None, string? reason = null)
+    public async Task AddOverwriteAsync(DiscordRole role, DiscordPermissions allow = default, DiscordPermissions deny = default, string? reason = null)
         => await this.Discord.ApiClient.EditChannelPermissionsAsync(this.Id, role.Id, allow, deny, "role", reason);
 
     /// <summary>
@@ -771,7 +813,7 @@ public class DiscordChannel : SnowflakeObject, IEquatable<DiscordChannel>
     /// <param name="member">The member to have the permission deleted.</param>
     /// <param name="reason">Reason for audit logs.</param>
     /// <returns></returns>
-    /// <exception cref="UnauthorizedException">Thrown when the client does not have the <see cref="DiscordPermissions.ManageRoles"/> permission.</exception>
+    /// <exception cref="UnauthorizedException">Thrown when the client does not have the <see cref="DiscordPermission.ManageRoles"/> permission.</exception>
     /// <exception cref="NotFoundException">Thrown when the channel does not exist.</exception>
     /// <exception cref="BadRequestException">Thrown when an invalid parameter was provided.</exception>
     /// <exception cref="ServerErrorException">Thrown when Discord is unable to process the request.</exception>
@@ -784,7 +826,7 @@ public class DiscordChannel : SnowflakeObject, IEquatable<DiscordChannel>
     /// <param name="role">The role to have the permission deleted.</param>
     /// <param name="reason">Reason for audit logs.</param>
     /// <returns></returns>
-    /// <exception cref="UnauthorizedException">Thrown when the client does not have the <see cref="DiscordPermissions.ManageRoles"/> permission.</exception>
+    /// <exception cref="UnauthorizedException">Thrown when the client does not have the <see cref="DiscordPermission.ManageRoles"/> permission.</exception>
     /// <exception cref="NotFoundException">Thrown when the channel does not exist.</exception>
     /// <exception cref="BadRequestException">Thrown when an invalid parameter was provided.</exception>
     /// <exception cref="ServerErrorException">Thrown when Discord is unable to process the request.</exception>
@@ -814,7 +856,7 @@ public class DiscordChannel : SnowflakeObject, IEquatable<DiscordChannel>
     /// Returns all pinned messages
     /// </summary>
     /// <returns></returns>
-    /// <exception cref="UnauthorizedException">Thrown when the client does not have the <see cref="DiscordPermissions.AccessChannels"/> permission.</exception>
+    /// <exception cref="UnauthorizedException">Thrown when the client does not have the <see cref="DiscordPermission.ViewChannel"/> permission.</exception>
     /// <exception cref="NotFoundException">Thrown when the channel does not exist.</exception>
     /// <exception cref="BadRequestException">Thrown when an invalid parameter was provided.</exception>
     /// <exception cref="ServerErrorException">Thrown when Discord is unable to process the request.</exception>
@@ -829,7 +871,7 @@ public class DiscordChannel : SnowflakeObject, IEquatable<DiscordChannel>
     /// <param name="avatar">The image for the default webhook avatar.</param>
     /// <param name="reason">Reason for audit logs.</param>
     /// <returns></returns>
-    /// <exception cref="UnauthorizedException">Thrown when the client does not have the <see cref="DiscordPermissions.ManageWebhooks"/> permission.</exception>
+    /// <exception cref="UnauthorizedException">Thrown when the client does not have the <see cref="DiscordPermission.ManageWebhooks"/> permission.</exception>
     /// <exception cref="NotFoundException">Thrown when the channel does not exist.</exception>
     /// <exception cref="BadRequestException">Thrown when an invalid parameter was provided.</exception>
     /// <exception cref="ServerErrorException">Thrown when Discord is unable to process the request.</exception>
@@ -853,7 +895,7 @@ public class DiscordChannel : SnowflakeObject, IEquatable<DiscordChannel>
     /// Returns a list of webhooks
     /// </summary>
     /// <returns></returns>
-    /// <exception cref="UnauthorizedException">Thrown when the client does not have the <see cref="DiscordPermissions.ManageWebhooks"/> permission.</exception>
+    /// <exception cref="UnauthorizedException">Thrown when the client does not have the <see cref="DiscordPermission.ManageWebhooks"/> permission.</exception>
     /// <exception cref="NotFoundException">Thrown when the channel does not exist.</exception>
     /// <exception cref="ServerErrorException">Thrown when Discord is unable to process the request.</exception>
     public async Task<IReadOnlyList<DiscordWebhook>> GetWebhooksAsync()
@@ -864,7 +906,7 @@ public class DiscordChannel : SnowflakeObject, IEquatable<DiscordChannel>
     /// </summary>
     /// <param name="member">The member to be moved.</param>
     /// <returns></returns>
-    /// <exception cref="UnauthorizedException">Thrown when the client does not have the <see cref="DiscordPermissions.MoveMembers"/> permission.</exception>
+    /// <exception cref="UnauthorizedException">Thrown when the client does not have the <see cref="DiscordPermission.MoveMembers"/> permission.</exception>
     /// <exception cref="NotFoundException">Thrown when the channel does not exists or if the Member does not exists.</exception>
     /// <exception cref="BadRequestException">Thrown when an invalid parameter was provided.</exception>
     /// <exception cref="ServerErrorException">Thrown when Discord is unable to process the request.</exception>
@@ -884,7 +926,7 @@ public class DiscordChannel : SnowflakeObject, IEquatable<DiscordChannel>
     /// </summary>
     /// <param name="targetChannel">Channel to crosspost messages to</param>
     /// <exception cref="ArgumentException">Thrown when trying to follow a non-news channel</exception>
-    /// <exception cref="UnauthorizedException">Thrown when the current user doesn't have <see cref="DiscordPermissions.ManageWebhooks"/> on the target channel</exception>
+    /// <exception cref="UnauthorizedException">Thrown when the current user doesn't have <see cref="DiscordPermission.ManageWebhooks"/> on the target channel</exception>
     public async Task<DiscordFollowedChannel> FollowAsync(DiscordChannel targetChannel) => this.Type != DiscordChannelType.News
             ? throw new ArgumentException("Cannot follow a non-news channel.")
             : await this.Discord.ApiClient.FollowChannelAsync(this.Id, targetChannel.Id);
@@ -895,7 +937,7 @@ public class DiscordChannel : SnowflakeObject, IEquatable<DiscordChannel>
     /// <param name="message">Message to publish</param>
     /// <exception cref="ArgumentException">Thrown when the message has already been crossposted</exception>
     /// <exception cref="UnauthorizedException">
-    ///     Thrown when the current user doesn't have <see cref="DiscordPermissions.ManageWebhooks"/> and/or <see cref="DiscordPermissions.SendMessages"/>
+    ///     Thrown when the current user doesn't have <see cref="DiscordPermission.ManageWebhooks"/> and/or <see cref="DiscordPermission.SendMessages"/>
     /// </exception>
     public async Task<DiscordMessage> CrosspostMessageAsync(DiscordMessage message) => (message.Flags & DiscordMessageFlags.Crossposted) == DiscordMessageFlags.Crossposted
             ? throw new ArgumentException("Message is already crossposted.")
@@ -1011,7 +1053,7 @@ public class DiscordChannel : SnowflakeObject, IEquatable<DiscordChannel>
         perms |= mbRoles.Aggregate(DiscordPermissions.None, (c, role) => c | role.Permissions);
 
         // Administrator grants all permissions and cannot be overridden
-        if ((perms & DiscordPermissions.Administrator) == DiscordPermissions.Administrator)
+        if (perms.HasPermission(DiscordPermission.Administrator))
         {
             return DiscordPermissions.All;
         }
@@ -1081,7 +1123,7 @@ public class DiscordChannel : SnowflakeObject, IEquatable<DiscordChannel>
         perms |= role.Permissions;
 
         // Administrator grants all permissions and cannot be overridden
-        if ((perms & DiscordPermissions.Administrator) == DiscordPermissions.Administrator)
+        if (perms.HasPermission(DiscordPermission.Administrator))
         {
             return DiscordPermissions.All;
         }
