@@ -1,380 +1,320 @@
+#pragma warning disable IDE0040
+
 using System;
+using System.Buffers.Binary;
+using System.Collections.Generic;
+using System.Diagnostics;
+using System.Diagnostics.CodeAnalysis;
+using System.Globalization;
+using System.Numerics;
+using System.Runtime.CompilerServices;
+using System.Runtime.InteropServices;
+using System.Runtime.Intrinsics;
+using System.Text;
+
+using CommunityToolkit.HighPerformance.Helpers;
+
+using DSharpPlus.Net.Serialization;
+
+using Newtonsoft.Json;
+
+using HashCode = CommunityToolkit.HighPerformance.Helpers.HashCode<uint>;
 
 namespace DSharpPlus.Entities;
 
-public static class PermissionMethods
-{
-    /// <summary>
-    /// Calculates whether this permission set contains the given permission.
-    /// </summary>
-    /// <param name="p">The permissions to calculate from</param>
-    /// <param name="permission">permission you want to check</param>
-    /// <returns></returns>
-    public static bool HasPermission(this DiscordPermissions p, DiscordPermissions permission)
-        => p.HasFlag(DiscordPermissions.Administrator) || (p & permission) == permission;
-
-    /// <summary>
-    /// Grants permissions.
-    /// </summary>
-    /// <param name="p">The permissions to add to.</param>
-    /// <param name="grant">Permission to add.</param>
-    /// <returns></returns>
-    public static DiscordPermissions Grant(this DiscordPermissions p, DiscordPermissions grant) => p | grant;
-
-    /// <summary>
-    /// Revokes permissions.
-    /// </summary>
-    /// <param name="p">The permissions to take from.</param>
-    /// <param name="revoke">Permission to take.</param>
-    /// <returns></returns>
-    public static DiscordPermissions Revoke(this DiscordPermissions p, DiscordPermissions revoke) => p & ~revoke;
-}
-
 /// <summary>
-/// Whether a permission is allowed, denied or unset
+/// Represents a set of Discord permissions.
 /// </summary>
-public enum DiscordPermissionLevel
+[JsonConverter(typeof(DiscordPermissionsAsStringJsonConverter))]
+public readonly partial struct DiscordPermissions
+    : IEquatable<DiscordPermissions>
 {
-    /// <summary>
-    /// Said permission is Allowed
-    /// </summary>
-    Allowed,
+    // only change ContainerWidth here, the other two constants are automatically updated for internal uses
+    // for ContainerWidth, 1 width == 128 bits.
+    private const int ContainerWidth = 1;
+    private const int ContainerElementCount = ContainerWidth * 4;
+    private const int ContainerByteCount = ContainerWidth * 16;
+
+    private readonly DiscordPermissionContainer data;
 
     /// <summary>
-    /// Said permission is Denied
+    /// Creates a new instance of this type from exactly the specified permission.
     /// </summary>
-    Denied,
+    public DiscordPermissions(DiscordPermission permission)
+        => this.data.SetFlag((int)permission, true);
 
     /// <summary>
-    /// Said permission is Unset
+    /// Creates a new instance of this type from the specified permissions.
     /// </summary>
-    Unset
+    [OverloadResolutionPriority(1)]
+    public DiscordPermissions(params ReadOnlySpan<DiscordPermission> permissions)
+    {
+        foreach (DiscordPermission permission in permissions)
+        {
+            this.data.SetFlag((int)permission, true);
+        }
+    }
+
+    /// <summary>
+    /// Creates a new instance of this type from the specified permissions.
+    /// </summary>
+    [OverloadResolutionPriority(0)]
+    public DiscordPermissions(params IReadOnlyList<DiscordPermission> permissions)
+    {
+        foreach (DiscordPermission permission in permissions)
+        {
+            this.data.SetFlag((int)permission, true);
+        }
+    }
+
+    /// <summary>
+    /// Creates a new instance of this type from the specified big integer. This assumes that the data is unsigned.
+    /// </summary>
+    public DiscordPermissions(BigInteger permissionSet)
+    {
+        Span<byte> buffer = MemoryMarshal.Cast<uint, byte>(MemoryMarshal.CreateSpan(ref this.data[0], ContainerElementCount));
+
+        if (!permissionSet.TryWriteBytes(buffer, out _, isUnsigned: true))
+        {
+            // we don't want to fail in release mode, which would break perfectly working code because the library
+            // hasn't been updated to support a new permission or because Discord is testing in prod again.
+            // seeing this assertion in dev should be an indication to expand this type.
+            Debug.Assert(false, "The amount of permissions DSharpPlus can represent has been exceeded.");
+        }
+    }
+
+    /// <summary>
+    /// Creates a new instance of this type from the specified raw data. This assumes that the data is unsigned.
+    /// </summary>
+    public DiscordPermissions(scoped ReadOnlySpan<byte> raw)
+    {
+        Span<byte> buffer = MemoryMarshal.Cast<uint, byte>(MemoryMarshal.CreateSpan(ref this.data[0], ContainerElementCount * 4));
+
+        if (!raw.TryCopyTo(buffer))
+        {
+            // we don't want to fail in release mode, which would break perfectly working code because the library
+            // hasn't been updated to support a new permission or because Discord is testing in prod again.
+            // seeing this assertion in dev should be an indication to expand this type.
+            Debug.Assert(false, "The amount of permissions DSharpPlus can represent has been exceeded.");
+        }
+    }
+
+    /// <summary>
+    /// A copy constructor that sets an arbitrary amount of flags to their respective values.
+    /// </summary>
+    private DiscordPermissions
+    (
+        scoped ReadOnlySpan<byte> raw,
+        ReadOnlySpan<DiscordPermission> setPermissions,
+        ReadOnlySpan<DiscordPermission> removePermissions
+    )
+        : this(raw)
+    {
+        foreach (DiscordPermission permission in setPermissions)
+        {
+            this.data.SetFlag((int)permission, true);
+        }
+
+        foreach (DiscordPermission permission in removePermissions)
+        {
+            this.data.SetFlag((int)permission, false);
+        }
+    }
+
+    /// <summary>
+    /// A copy constructor that sets one specific flag to the specified value.
+    /// </summary>
+    private DiscordPermissions(DiscordPermissions original, int index, bool flag)
+    {
+        this.data = original.data;
+        this.data.SetFlag(index, flag);
+    }
+
+    public static implicit operator DiscordPermissions(DiscordPermission initial) => new(initial);
+
+    /// <summary>
+    /// Returns an empty Discord permission set.
+    /// </summary>
+    public static DiscordPermissions None => default;
+
+    /// <summary>
+    /// Returns a full Discord permission set with all flags set to true.
+    /// </summary>
+    public static DiscordPermissions AllBitsSet
+    {
+        get
+        {
+            Span<byte> result = stackalloc byte[ContainerByteCount];
+
+            for (int i = 0; i < ContainerElementCount; i += 16)
+            {
+                Vector128.StoreUnsafe(Vector128<byte>.AllBitsSet, ref result[i]);
+            }
+
+            return new(result);
+        }
+    }
+
+    /// <summary>
+    /// Returns a Discord permission set with all documented permissions set to true.
+    /// </summary>
+    public static DiscordPermissions All { get; } = new(DiscordPermissionExtensions.GetValues());
+
+    [UnscopedRef]
+    private ReadOnlySpan<byte> AsSpan
+        => MemoryMarshal.Cast<uint, byte>((ReadOnlySpan<uint>)this.data);
+
+    private bool GetFlag(int index)
+        => this.data.HasFlag(index);
+
+    /// <summary>
+    /// Determines whether this Discord permission set is equal to the provided object.
+    /// </summary>
+    public override bool Equals([NotNullWhen(true)] object? obj)
+        => obj is DiscordPermissions permissions && Equals(permissions);
+
+    /// <summary>
+    /// Determines whether this Discord permission set is equal to the provided Discord permission set.
+    /// </summary>
+    public bool Equals(DiscordPermissions other)
+    {
+        for (int i = 0; i < ContainerElementCount; i += 4)
+        {
+            Vector128<uint> current = Vector128.LoadUnsafe(in this.data[i]);
+            Vector128<uint> comparison = Vector128.LoadUnsafe(in other.data[i]);
+
+            if (current != comparison)
+            {
+                return false;
+            }
+        }
+
+        return true;
+    }
+
+    /// <summary>
+    /// Returns a string representation of this permission set.
+    /// </summary>
+    public override string ToString() => ToString("a placeholder format string that doesn't do anything");
+
+    /// <summary>
+    /// Returns a string representation of this permission set, according to the provided format string.
+    /// </summary>
+    /// <param name="format">
+    /// Specifies the format in which the string should be created. Currently supported formats are: <br/>
+    /// - <c>raw</c>: This prints the raw, byte-wise backing data of this instance. <br/>
+    /// - <c>name</c>: This prints each flag by name, separated by commas. <br/>
+    /// - anything else will print the integer value contained in this <see cref="DiscordPermissions"/> instance.
+    /// </param>
+    public string ToString(string format)
+    {
+        if (format == "raw")
+        {
+            StringBuilder builder = new("DiscordPermissions - raw value:");
+
+            foreach (byte b in this.AsSpan)
+            {
+                _ = builder.Append(' ');
+                _ = builder.Append(b.ToString("x2", CultureInfo.InvariantCulture));
+            }
+
+            return builder.ToString();
+        }
+        else if (format == "name")
+        {
+            int pop = 0;
+
+            for (int i = 0; i < ContainerElementCount; i += 4)
+            {
+                pop += BitOperations.PopCount(this.data[i]);
+                pop += BitOperations.PopCount(this.data[i + 1]);
+                pop += BitOperations.PopCount(this.data[i + 2]);
+                pop += BitOperations.PopCount(this.data[i + 3]);
+            }
+
+            string[] names = new string[pop];
+            DiscordPermissionEnumerator enumerator = new(this.data);
+
+            for (int i = 0; i < pop; i++)
+            {
+                _ = enumerator.MoveNext();
+                names[i] = enumerator.Current.ToStringFast();
+            }
+
+            return string.Join(", ", names);
+        }
+        else
+        {
+            Span<byte> buffer = stackalloc byte[ContainerElementCount * 4];
+            this.AsSpan.CopyTo(buffer);
+
+            if (!BitConverter.IsLittleEndian)
+            {
+                Span<uint> bigEndianWorkaround = MemoryMarshal.Cast<byte, uint>(buffer);
+                BinaryPrimitives.ReverseEndianness(bigEndianWorkaround, bigEndianWorkaround);
+            }
+
+            return new BigInteger(buffer, true, false).ToString(CultureInfo.InvariantCulture);
+        }
+    }
+
+    /// <summary>
+    /// Calculates a hash code for this Discord permission set. The hash code is only guaranteed to be consistent
+    /// within a process, and sharing this data across process boundaries is dangerous.
+    /// </summary>
+    public override int GetHashCode()
+        => HashCode.Combine(this.data);
+
+    /// <summary>
+    /// Provides an enumeration of all permissions specified by this set.
+    /// </summary>
+    public DiscordPermissionEnumerable EnumeratePermissions()
+        => new(this.data);
+
+    public static bool operator ==(DiscordPermissions left, DiscordPermissions right) => left.Equals(right);
+    public static bool operator !=(DiscordPermissions left, DiscordPermissions right) => !(left == right);
+
+    // we will be using an inline array from the start here so that further increases in the bit width
+    // only require increasing this number instead of switching to a new backing implementation strategy.
+    // if Discord changes the way permissions are represented in the API, this will obviously have to change.
+    //
+    // this should always be backed by a 32-bit integer, to make our life easier around popcnt and BitHelper.
+    //
+    /// <summary>
+    /// Represents a container for the backing storage of Discord permissions.
+    /// </summary>
+    [InlineArray(ContainerElementCount)]
+    internal struct DiscordPermissionContainer
+    {
+        public uint value;
+
+        /// <summary>
+        /// Sets a specified flag to the specific value. This function fails in debug mode if the flag was out of range.
+        /// </summary>
+        public void SetFlag(int index, bool value)
+        {
+            int fieldIndex = index >> 5;
+
+            Debug.Assert(fieldIndex < ContainerElementCount);
+
+            int bitIndex = index & 0x1F;
+            ref uint segment = ref this[fieldIndex];
+            BitHelper.SetFlag(ref segment, bitIndex, value);
+        }
+
+        /// <summary>
+        /// Returns the value of a specified flag. This function fails in debug mode if the flag was out of range.
+        /// </summary>
+        public readonly bool HasFlag(int index)
+        {
+            int fieldIndex = index >> 5;
+
+            Debug.Assert(fieldIndex < ContainerElementCount);
+
+            int bitIndex = index & 0x1F;
+            uint segment = this[fieldIndex];
+            return BitHelper.HasFlag(segment, bitIndex);
+        }
+    }
 }
 
-/// <summary>
-/// Bitwise permission flags.
-/// </summary>
-[Flags]
-public enum DiscordPermissions : long
-{
-    /// <summary>
-    /// Indicates no permissions given.
-    /// </summary>
-    [PermissionString("No permissions")]
-    None = 0x0000000000000000,
-
-    /// <summary>
-    /// Allows creation of instant channel invites.
-    /// </summary>
-    [PermissionString("Create instant invites")]
-    CreateInstantInvite = 0x0000000000000001,
-
-    /// <summary>
-    /// Allows kicking members.
-    /// </summary>
-    [PermissionString("Kick members")]
-    KickMembers = 0x0000000000000002,
-
-    /// <summary>
-    /// Allows banning and unbanning members.
-    /// </summary>
-    [PermissionString("Ban members")]
-    BanMembers = 0x0000000000000004,
-
-    /// <summary>
-    /// Enables full access on a given guild. This also overrides other permissions.
-    /// </summary>
-    [PermissionString("Administrator")]
-    Administrator = 0x0000000000000008,
-
-    /// <summary>
-    /// Allows managing channels.
-    /// </summary>
-    [PermissionString("Manage channels")]
-    ManageChannels = 0x0000000000000010,
-
-    /// <summary>
-    /// Allows managing the guild.
-    /// </summary>
-    [PermissionString("Manage guild")]
-    ManageGuild = 0x0000000000000020,
-
-    /// <summary>
-    /// Allows adding reactions to messages.
-    /// </summary>
-    [PermissionString("Add reactions")]
-    AddReactions = 0x0000000000000040,
-
-    /// <summary>
-    /// Allows viewing audit log entries.
-    /// </summary>
-    [PermissionString("View audit log")]
-    ViewAuditLog = 0x0000000000000080,
-
-    /// <summary>
-    /// Allows the use of priority speaker.
-    /// </summary>
-    [PermissionString("Use priority speaker")]
-    PrioritySpeaker = 0x0000000000000100,
-
-    /// <summary>
-    /// Allows accessing text and voice channels. Disabling this permission hides channels.
-    /// </summary>
-    [PermissionString("Read messages")]
-    AccessChannels = 0x0000000000000400,
-
-    /// <summary>
-    /// Allows sending messages (does not allow sending messages in threads).
-    /// </summary>
-    [PermissionString("Send messages")]
-    SendMessages = 0x0000000000000800,
-
-    /// <summary>
-    /// Allows sending text-to-speech messages.
-    /// </summary>
-    [PermissionString("Send TTS messages")]
-    SendTtsMessages = 0x0000000000001000,
-
-    /// <summary>
-    /// Allows managing messages of other users.
-    /// </summary>
-    [PermissionString("Manage messages")]
-    ManageMessages = 0x0000000000002000,
-
-    /// <summary>
-    /// Allows embedding content in messages.
-    /// </summary>
-    [PermissionString("Use embeds")]
-    EmbedLinks = 0x0000000000004000,
-
-    /// <summary>
-    /// Allows uploading files.
-    /// </summary>
-    [PermissionString("Attach files")]
-    AttachFiles = 0x0000000000008000,
-
-    /// <summary>
-    /// Allows reading message history.
-    /// </summary>
-    [PermissionString("Read message history")]
-    ReadMessageHistory = 0x0000000000010000,
-
-    /// <summary>
-    /// Allows using @everyone and @here mentions.
-    /// </summary>
-    [PermissionString("Mention everyone")]
-    MentionEveryone = 0x0000000000020000,
-
-    /// <summary>
-    /// Allows using emojis or stickers from external servers, such as twitch or nitro emojis.
-    /// </summary>
-    [PermissionString("Use external emojis and stickers")]
-    UseExternalEmojis = 0x0000000000040000,
-
-    /// <summary>
-    /// Allows connecting to voice chat.
-    /// </summary>
-    [PermissionString("Use voice chat")]
-    UseVoice = 0x0000000000100000,
-
-    /// <summary>
-    /// Allows speaking in voice chat.
-    /// </summary>
-    [PermissionString("Speak")]
-    Speak = 0x0000000000200000,
-
-    /// <summary>
-    /// Allows muting other members in voice chat.
-    /// </summary>
-    [PermissionString("Mute voice chat members")]
-    MuteMembers = 0x0000000000400000,
-
-    /// <summary>
-    /// Allows deafening other members in voice chat.
-    /// </summary>
-    [PermissionString("Deafen voice chat members")]
-    DeafenMembers = 0x0000000000800000,
-
-    /// <summary>
-    /// Allows moving voice chat members.
-    /// </summary>
-    [PermissionString("Move voice chat members")]
-    MoveMembers = 0x0000000001000000,
-
-    /// <summary>
-    /// Allows using voice activation in voice chat. Revoking this will usage of push-to-talk.
-    /// </summary>
-    [PermissionString("Use voice activity detection")]
-    UseVoiceDetection = 0x0000000002000000,
-
-    /// <summary>
-    /// Allows changing of own nickname.
-    /// </summary>
-    [PermissionString("Change own nickname")]
-    ChangeNickname = 0x0000000004000000,
-
-    /// <summary>
-    /// Allows managing nicknames of other members.
-    /// </summary>
-    [PermissionString("Manage nicknames")]
-    ManageNicknames = 0x0000000008000000,
-
-    /// <summary>
-    /// Allows managing roles in a guild.
-    /// </summary>
-    [PermissionString("Manage roles")]
-    ManageRoles = 0x0000000010000000,
-
-    /// <summary>
-    /// Allows managing webhooks in a guild.
-    /// </summary>
-    [PermissionString("Manage webhooks")]
-    ManageWebhooks = 0x0000000020000000,
-
-    /// <summary>
-    /// Allows managing guild emoji and stickers.
-    /// </summary>
-    [PermissionString("Manage emoji and stickers")]
-    ManageEmojis = 0x0000000040000000,
-
-    /// <summary>
-    /// Allows the user to go live.
-    /// </summary>
-    [PermissionString("Allow stream")]
-    Stream = 0x0000000000000200,
-
-    /// <summary>
-    /// Allows the user to use application commands.
-    /// </summary>
-    [PermissionString("Use application commands")]
-    UseApplicationCommands = 0x0000000080000000,
-
-    /// <summary>
-    /// Allows for requesting to speak in stage channels.
-    /// </summary>
-    [PermissionString("Request to speak")]
-    RequestToSpeak = 0x0000000100000000,
-
-    /// <summary>
-    /// Allows for managing scheduled guild events.
-    /// </summary>
-    [PermissionString("Manage Scheduled Events")]
-    ManageEvents = 0x0000000200000000,
-
-    /// <summary>
-    /// Allows for deleting and archiving threads, and viewing all private threads.
-    /// </summary>
-    [PermissionString("Manage Threads")]
-    ManageThreads = 0x0000000400000000,
-
-    /// <summary>
-    /// Allows for creating public threads.
-    /// </summary>
-    [PermissionString("Create Public Threads")]
-    CreatePublicThreads = 0x0000000800000000,
-
-    /// <summary>
-    /// Allows for creating private threads.
-    /// </summary>
-    [PermissionString("Create Private Threads")]
-    CreatePrivateThreads = 0x0000001000000000,
-
-    /// <summary>
-    /// Allows the usage of custom stickers from other servers.
-    /// </summary>
-    [PermissionString("Use external Stickers")]
-    UseExternalStickers = 0x0000002000000000,
-
-    /// <summary>
-    /// Allows for sending messages in threads.
-    /// </summary>
-    [PermissionString("Send messages in Threads")]
-    SendMessagesInThreads = 0x0000004000000000,
-
-    /// <summary>
-    /// Allows for launching activities (applications with the `EMBEDDED` flag) in a voice channel.
-    /// </summary>
-    [PermissionString("Start Embedded Activities")]
-    StartEmbeddedActivities = 0x0000008000000000,
-
-    /// <summary>
-    /// Allows for moderating (Timeout) members in a guild.
-    /// </summary>
-    [PermissionString("Moderate Members")]
-    ModerateMembers = 0x0000010000000000,
-
-    /// <summary>
-    /// Allows for viewing role subscription insights.
-    /// </summary>
-    [PermissionString("View Creator Monetization Analytics")]
-    ViewCreatorMonetizationAnalytics = 0x0000020000000000,
-
-    /// <summary>
-    /// Allows for using soundboard in a voice channel.
-    /// </summary>
-    [PermissionString("Use Soundboard")]
-    UseSoundboard = 0x0000040000000000,
-
-    /// <summary>
-    /// Allows for creating emojis, stickers, soundboard sounds, editing and deleting those created by the current user.
-    /// </summary>
-    [PermissionString("Create Guild Expressions")]
-    CreateGuildExpressions = 0x0000080000000000,
-
-    /// <summary>
-    /// Allows for creating scheduled events, editing and deleting those created by the current user.
-    /// </summary>
-    [PermissionString("Create Events")]
-    CreateEvents = 0x0000100000000000,
-
-    /// <summary>
-    /// Allows the usage of custom soundboard sounds from other servers.
-    /// </summary>
-    [PermissionString("Use External Sounds")]
-    UseExternalSounds = 0x0000200000000000,
-
-    /// <summary>
-    /// Allows sending voice messages.
-    /// </summary>
-    [PermissionString("Send Voice Messages")]
-    SendVoiceMessages = 0x0000400000000000,
-
-    /// <summary>
-    /// Allows sending polls.
-    /// </summary>
-    [PermissionString("Send Polls")]
-    SendPolls = 0x0002000000000000,
-
-    /// <summary>
-    /// Allows user-installed apps to send public responses.
-    /// When disabled, users will still be allowed to use their apps but the responses will be ephemeral.
-    /// This only applies to apps not also installed to the server.
-    /// </summary>
-    [PermissionString("Use External Apps")]
-    UseExternalApps = 0x0004000000000000,
-
-    /// <summary>
-    /// Indicates all permissions are granted
-    /// </summary>
-    [PermissionString("All permissions")]
-    // TODO: When Discord adds a new permission, update this and prepend the new permission.
-    All = UseExternalApps | SendPolls | SendVoiceMessages | UseExternalSounds | CreateEvents | CreateGuildExpressions | UseSoundboard | ViewCreatorMonetizationAnalytics | ModerateMembers | StartEmbeddedActivities | SendMessagesInThreads | UseExternalStickers | CreatePrivateThreads | CreatePublicThreads | ManageThreads | ManageEvents | RequestToSpeak | UseApplicationCommands | ManageEmojis | ManageWebhooks | ManageRoles | ManageNicknames | ChangeNickname | UseVoiceDetection | MoveMembers | DeafenMembers | MuteMembers | Speak | UseVoice | UseExternalEmojis | MentionEveryone | ReadMessageHistory | AttachFiles | EmbedLinks | ManageMessages | SendTtsMessages | SendMessages | AccessChannels | Stream | PrioritySpeaker | ViewAuditLog | AddReactions | ManageGuild | ManageChannels | Administrator | BanMembers | KickMembers | CreateInstantInvite,
-}
-
-/// <summary>
-/// Defines a readable name for this permission.
-/// </summary>
-[AttributeUsage(AttributeTargets.Field, AllowMultiple = false)]
-public sealed class PermissionStringAttribute : Attribute
-{
-    /// <summary>
-    /// Gets the readable name for this permission.
-    /// </summary>
-    public string String { get; }
-
-    /// <summary>
-    /// Defines a readable name for this permission.
-    /// </summary>
-    /// <param name="str">Readable name for this permission.</param>
-    public PermissionStringAttribute(string str) => this.String = str;
-}
