@@ -8,7 +8,7 @@ using System.Threading.Tasks;
 using CommunityToolkit.HighPerformance.Buffers;
 
 using DSharpPlus.Logging;
-using DSharpPlus.Net.WebSocket;
+using DSharpPlus.Net.Gateway.Compression;
 
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
@@ -22,7 +22,7 @@ internal sealed class TransportService : ITransportService
     private ClientWebSocket socket;
     private readonly ArrayPoolBufferWriter<byte> writer;
     private readonly ArrayPoolBufferWriter<byte> decompressedWriter;
-    private readonly PayloadDecompressor decompressor;
+    private readonly IPayloadDecompressor decompressor;
     private readonly ILoggerFactory factory;
 
     private readonly bool streamingDeserialization;
@@ -33,7 +33,7 @@ internal sealed class TransportService : ITransportService
     public TransportService
     (
         ILoggerFactory factory,
-        PayloadDecompressor decompressor,
+        IPayloadDecompressor decompressor,
         IOptions<GatewayClientOptions> options
     )
     {
@@ -55,6 +55,7 @@ internal sealed class TransportService : ITransportService
             : this.factory.CreateLogger($"DSharpPlus.Net.Gateway.ITransportService - Shard {shardId}");
 
         this.socket = new();
+        this.decompressor.Initialize();
 
         ObjectDisposedException.ThrowIf(this.isDisposed, this);
 
@@ -66,10 +67,10 @@ internal sealed class TransportService : ITransportService
 
         this.logger.LogTrace("Connecting to the Discord gateway.");
 
-        await this.socket.ConnectAsync(new($"{url}?v=10&encoding=json"), CancellationToken.None);
+        await this.socket.ConnectAsync(new(url), CancellationToken.None);
         this.isConnected = true;
 
-        this.logger.LogTrace("Connected to the Discord websocket.");
+        this.logger.LogDebug("Connected to the Discord websocket, using {compression} compression.", this.decompressor.Name);
     }
 
     /// <inheritdoc/>
@@ -125,6 +126,7 @@ internal sealed class TransportService : ITransportService
                 break;
         }
 
+        this.decompressor.Reset();
         this.socket.Dispose();
     }
 
