@@ -6,7 +6,7 @@ namespace DSharpPlus.SlashCommands;
 [AttributeUsage(AttributeTargets.Method | AttributeTargets.Class)]
 public class SlashCommandPermissionsAttribute : Attribute
 {
-    public DiscordPermissions Permissions { get; }
+    public DiscordPermission[] Permissions { get; }
 
-    public SlashCommandPermissionsAttribute(DiscordPermissions permissions) => this.Permissions = permissions;
+    public SlashCommandPermissionsAttribute(params DiscordPermission[] permissions) => this.Permissions = permissions;
 }
